requirements.txt
discord.py
aiohttp==3.4.4
python-dotenv==0.10.1
yarl==1.3.0
multidict==4.5.2
attrs==18.2.0
black
